+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Upcoming
+
+Refactor.
+
 # v0.13.0
 
 ## Non-breaking